+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+unreleased
+ - valid predit-fn results
+ - new ns scicloj.metamorph.ml.column-metric with metric-fns operating on datasets, not seq-of-numbers
+
 1.3.2
  - removed clj-memory-meter from src-code 
 1.3.1